third_party/WebKit/LayoutTests/virtual/outofblink-cors/http/tests/xmlhttprequest/redirect-cross-origin-sync-double-expected.txt

CONSOLE WARNING: line 25: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 26: Failed to load http://localhost:8000/xmlhttprequest/resources/redirect.php?url=/xmlhttprequest/resources/reply.xml: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Test that a cross-origin chain of redirects to a server that responds is indistinguishable from one that does not. Should say PASS:

PASS
